+++ application/views/peminjaman/sales/index.php
@@ -1,3 +1,7 @@
+<?php
+$user = $this->session->userdata();
+?>
+
 <script>
   function deleteConfirm(url) {
     $('#btn-delete').attr('href', url);
@@ -18,8 +22,11 @@
   <div class="card shadow mb-4">
     <div class="card-header py-3">
       <h6 class="m-0 font-weight-bold text-primary float-left"><?= $title; ?></h6>
-
-      <h6 class="m-0 font-weight-bold text-primary float-right"><a href="<?= site_url('/peminjaman/add')?>"><i class="fas fa-plus"></i> Tambah Peminjaman</a></h6>
+      <h6 class="m-0 font-weight-bold text-primary float-right">
+        <?php if (in_array($user['role_id'], array(1, 2))) { ?>
+          <a href="<?= site_url('/peminjaman/add') ?>"><i class="fas fa-plus"></i> Tambah Peminjaman</a>
+        <?php } ?>
+      </h6>
     </div>
     <div class="card-body">
       <div class="table-responsive">
@@ -33,6 +40,7 @@
               <th>Tanggal</th>
               <th>Closing Date</th>
               <th>Catatan</th>
+              <th>Status</th>
               <th>Action</th>
             </tr>
           </thead>
@@ -48,15 +56,28 @@
                   <td><?= $item['closingdate']; ?></td>
                   <td><?= $item['note']; ?></td>
                   <td>
+                    <?php if ($item['status'] == "PENDING") echo ('<p class="badge badge-warning">' . $item['status'] . '</p>');
+                    else if ($item['status'] == "PROCESS") echo ('<p class="badge badge-primary">' . $item['status'] . '</p>');
+                    else if ($item['status'] == "SUCCESS") echo ('<p class="badge badge-success">' . $item['status'] . '</p>');
+                    else echo ('<p class="badge badge-danger">' . $item['status'] . '</p>');
+                    ?>
+                  </td>
+                  <td>
                     <a class="badge badge-primary" style="font-size:14px;" href="<?= site_url('peminjaman/detail/' . $item['id_peminjaman']); ?>"><i class="fas fa fa-eye"></i> Detail</a>
-                    <a class="badge badge-success" style="font-size:14px;" href="<?= site_url('peminjaman/edit/' . $item['id_peminjaman']); ?>"><i class="fas fa fa-pen"></i> Perbarui</a>
-                    <a class="badge badge-danger" style="font-size:14px;" href="#!" onclick="deleteConfirm('<?= site_url('peminjaman/delete/' . $item['id_peminjaman']); ?>')"><i class="fas fa fa-trash"></i> Hapus</a>
+                    <?php if (in_array($user['role_id'], array(1, 2, 3))) { ?>
+                      <?php if ($item['status'] == "PENDING" && $user['role_id'] == 2 || ($item['status'] == "PENDING" || $item['status'] == "PROCESS") && in_array($user['role_id'], array(3, 8))) { ?>
+                        <a class="badge badge-success" style="font-size:14px;" href="<?= site_url('peminjaman/edit/' . $item['id_peminjaman']); ?>"><i class="fas fa fa-pen"></i> Perbarui</a>
+                      <?php } ?>
+                    <?php } ?>
+                    <?php if (in_array($user['role_id'], array(1, 2)) && $item['status'] == "PENDING") { ?>
+                      <a class="badge badge-danger" style="font-size:14px;" href="#!" onclick="deleteConfirm('<?= site_url('peminjaman/delete/' . $item['id_peminjaman']); ?>')"><i class="fas fa fa-trash"></i> Hapus</a>
+                    <?php } ?>
                   </td>
                 </tr>
               <?php endforeach; ?>
             <?php } else { ?>
               <tr>
-                <td colspan="8" class="text-center">Data Peminjaman Masih Kosong</td>
+                <td colspan="9" class="text-center">Data Peminjaman Masih Kosong</td>
               </tr>
             <?php } ?>
           </tbody>
